requirements.txt
numpy
scipy
matplotlib
tensorflow
pillow
scikit-image
efficientnet
pyswarm